refactor: モジュールの抽出
