+++ dev/20190508/k2works/package.json
@@ -7,8 +7,10 @@
     "start": "nf start -j Procfile",
     "build": "composer install",
     "doc": "markdown README.md > ./src/readme.html",
-    "test": "./vendor/bin/phpunit --bootstrap vendor/autoload.php src/index",
-    "watch": "watch 'npm-run-all test' ./src/",
+    "test": "./vendor/bin/phpunit --color --bootstrap vendor/autoload.php src/tests/FizzBuzzTest",
+    "watch": "npm-run-all watch:*",
+    "watch:src": "watch 'npm-run-all test' ./src",
+    "watch:test": "watch 'npm-run-all test' ./src/tests",
     "app:front": "browser-sync start --config bs-config.js",
     "app:server": "php -S 0.0.0.0:8000 -t src/"
   },
